basic sandwich
